[LLT-5552] Remove Teliod PID file
